misc tidy-up
